agent/pyproject.toml
[project]
name = "ai-over-bi-agent"
version = "0.1.0"
description = "AI over BI Agent Backend — Google ADK + Anthropic Claude + AG-UI + FastAPI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "google-adk>=0.5.0",
    "ag-ui-adk>=0.1.0",
    "litellm>=1.40.0",
    "anthropic>=0.40.0",
    "pydantic-settings>=2.3.0",
    "httpx>=0.27.0",
    "copilotkit>=0.1.31",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "mypy>=1.10.0",
    "ruff>=0.4.0",
    "black>=24.4.0",
    "pre-commit>=3.7.0",
]

[project.scripts]
ai-over-bi-serve = "ai_over_bi.main:main"
ai-over-bi-seed = "ai_over_bi.data.seed:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ai_over_bi"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
